% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRFLDA.R
\name{WRLDA}
\alias{WRLDA}
\title{MRF-LDA}
\usage{
WRLDA(
  corpus,
  k,
  type = "base",
  seed = 0,
  iterations = 1500,
  priors = NULL,
  lambda = NULL,
  beta_init = NULL,
  ldamodel = NULL,
  verbose = FALSE,
  neiPath = NULL,
  threshold = 1,
  Z = NULL
)
}
\description{
This function do the deconvolution with genes network.
}
